guards

+++ Sources/PSQLKit/Directives/From/DeleteDirective.swift
@@ -16,6 +16,7 @@ public struct DeleteDirective<T: FromSQLExpression>: SQLExpression {
     }
     
     public func serialize(to serializer: inout SQLSerializer) {
+        guard !content.fromIsNull else { return }
         serializer.write("DELETE FROM")
         serializer.writeSpace()
         content.fromSqlExpression.serialize(to: &serializer)
